+++ src/main/java/org/smoothbuild/lang/base/NativeWrapper.java
@@ -1,0 +1,7 @@
+package org.smoothbuild.lang.base;
+
+public interface NativeWrapper {
+  public String name();
+
+  public Native nativ();
+}