+++ src/main/java/org/smoothbuild/nativ/Natives.java
@@ -7,6 +7,7 @@ import java.lang.reflect.Method;
 import java.lang.reflect.Parameter;
 import java.util.List;
 import java.util.Map;
+import java.util.function.BiFunction;
 
 import org.smoothbuild.cli.console.Logger;
 import org.smoothbuild.db.record.base.Record;
@@ -14,8 +15,10 @@ import org.smoothbuild.lang.base.Native;
 import org.smoothbuild.lang.base.type.Type;
 import org.smoothbuild.lang.parse.ast.Ast;
 import org.smoothbuild.lang.parse.ast.AstVisitor;
+import org.smoothbuild.lang.parse.ast.EvaluableNode;
 import org.smoothbuild.lang.parse.ast.FuncNode;
 import org.smoothbuild.lang.parse.ast.ItemNode;
+import org.smoothbuild.lang.parse.ast.ValueNode;
 
 public class Natives {
   private final Map<String, Native> map;
@@ -29,40 +32,60 @@ public class Natives {
       @Override
       public void visitFunc(FuncNode func) {
         super.visitFunc(func);
-        if (map.containsKey(func.name())) {
-          Native nativ = map.get(func.name());
-          if (func.isNative()) {
-            assign(func, nativ);
+        assignNative(func, this::nativeParameterTypesMatchesFuncParameters);
+      }
+
+      @Override
+      public void visitValue(ValueNode value) {
+        super.visitValue(value);
+        assignNative(value, this::nativeHasOneParameter);
+      }
+
+      private void assignNative(EvaluableNode evaluable,
+          BiFunction<Native, EvaluableNode, Boolean> parameterChecker) {
+        if (map.containsKey(evaluable.name())) {
+          Native nativ = map.get(evaluable.name());
+          if (evaluable.isNative()) {
+            if (nativeTypesMatchesDeclaredTypes(evaluable, nativ, parameterChecker)) {
+              evaluable.setNative(nativ);
+            }
           } else {
-            logger.log(parseError(func, "Function '" + func.name()
+            logger.log(parseError(evaluable, "'" + evaluable.name()
                 + "' has both definition and native implementation in " + nativ.jarFile() + "."));
           }
         } else {
-          if (func.isNative()) {
-            logger.log(parseError(func, "Function '" + func.name()
+          if (evaluable.isNative()) {
+            logger.log(parseError(evaluable, "'" + evaluable.name()
                 + "' is native but does not have native implementation."));
           }
         }
       }
 
-      private void assign(FuncNode func, Native nativ) {
+      private boolean nativeTypesMatchesDeclaredTypes(EvaluableNode evaluable, Native nativ,
+          BiFunction<Native, EvaluableNode, Boolean> parameterChecker) {
         Method method = nativ.method();
-        Type resultType = func.type().get();
+        Type resultType = evaluable.type().get();
         Class<?> resultJType = method.getReturnType();
         if (!mapTypeToJType(resultType).equals(resultJType)) {
-          logger.log(parseError(func, "Function '" + func.name() + "' has result type "
-                  + resultType.q() + " so its native implementation result type must be "
-                  + mapTypeToJType(resultType).getCanonicalName() + " but it is "
-                  + resultJType.getCanonicalName() + "."));
-          return;
+          logger.log(parseError(evaluable, "'" + evaluable.name() + "' declares type "
+              + resultType.q() + " so its native implementation result type must be "
+              + mapTypeToJType(resultType).getCanonicalName() + " but it is "
+              + resultJType.getCanonicalName() + "."));
+          return false;
         }
-        Parameter[] nativeParams = method.getParameters();
+        return parameterChecker.apply(nativ, evaluable);
+      }
+
+      private boolean nativeParameterTypesMatchesFuncParameters(
+          Native nativ, EvaluableNode evaluable) {
+        FuncNode func = (FuncNode) evaluable;
+        Parameter[] nativeParams = nativ.method().getParameters();
         List<ItemNode> params = func.params();
         if (params.size() != nativeParams.length - 1) {
           logger.log(parseError(func, "Function '" + func.name() + "' has "
               + params.size() + " parameter(s) but its native implementation has "
               + (nativeParams.length - 1) + " parameter(s)."));
-          return;
+          return false;
         }
         for (int i = 0; i < params.size(); i++) {
           String declaredName = params.get(i).name();
@@ -76,10 +99,22 @@ public class Natives {
                 + paramType.name() + " so its native implementation type must be "
                 + expectedParamJType.getCanonicalName() + " but it is "
                 + paramJType.getCanonicalName() + "."));
-            return;
+            return false;
           }
         }
-        func.setNative(nativ);
+        return true;
+      }
+
+      private boolean nativeHasOneParameter(Native nativ, EvaluableNode evaluable) {
+        ValueNode value = (ValueNode) evaluable;
+        int paramCount = nativ.method().getParameters().length;
+        if (paramCount != 1) {
+          logger.log(parseError(value, "'" + value.name()
+              + "' has native implementation that has too many parameter(s) = " +
+              paramCount));
+          return false;
+        }
+        return true;
       }
     }.visitAst(ast);
   }